Two-Channel/requirements.txt
absl-py==2.1.0
alembic==1.13.3
anyio==4.5.2
astropy==5.2.2
cachetools==5.5.0
click==8.1.7
cloudpickle==3.1.0
colorlog==6.8.2
contourpy==1.1.1
dask==2023.5.0
distributed==2023.5.0
exceptiongroup==1.2.2
fsspec==2024.9.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
importlib-metadata==8.5.0
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.41.1
locket==1.0.0
mako==1.3.5
markdown==3.7
matplotlib==3.7.5
numba==0.58.1
oauthlib==3.2.2
opencv-python==4.4.0.46
optuna==4.0.0
packaging==24.1
pandas==2.0.3
partd==1.4.1
protobuf==5.29.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyerfa==2.0.0.3
pyparsing==3.1.4
python-telegram-bot==21.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.35
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.5.0
timm==0.4.12
toolz==1.0.0
torchaudio==2.4.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
wheel==0.44.0
yacs==0.1.8
zict==3.0.0
zipp==3.20.2
